requirements.txt
discord.py==2.7.1
aiohttp==3.13.3
openpyxl==3.1.5
PyMuPDF==1.27.2.2
pdfplumber==0.11.9
pillow==12.1.1
pytesseract==0.3.13
python-dotenv==1.2.2
requests==2.33.0